add read-only flag in transactional annotation

+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/program/notification/ProgramNotificationTemplateService.java
@@ -71,7 +71,7 @@ public interface ProgramNotificationTemplateService
 
     List<ProgramNotificationTemplate> getProgramStageLinkedToWebHookNotifications( ProgramStage programStage );
 
-    Long countProgramNotificationTemplates( ProgramNotificationTemplateParam programNotificationTemplateParam );
+    int countProgramNotificationTemplates( ProgramNotificationTemplateParam programNotificationTemplateParam );
 
     List<ProgramNotificationTemplate> getProgramNotificationTemplates(
         ProgramNotificationTemplateParam programNotificationTemplateParam );
